backend-python/requirements.txt
flask
mysql-connector-python
google-api-python-client
google-auth
